requirements.txt
python-dotenv==1.1.0
selenium==4.31.0
webdriver-manager==4.0.2
python-telegram-bot==22.0